added a categorical style that supports equidistant hue steps and optional data defined saturation.
the polygons border are rendered with a constant hue defined by the hazard level, while the polygon fill can be same as the border or the saturation can
data defined. 
In the latter case, the colors become brighter or darker (saturation) depending on their value in a field.

This allows creating maps that show a category and a graduation in the category