###Problem Wizard dialogues are not informative enough to allow users to make informed decisions on which EQ IF to use.
###Solution
Talk to @hadighasemi and @hyeukryu about how the three EQ IFs work.
Change text in dialogues to explain the differences between the IFs, and in what situations each IF should be used.

###Status
Have talked to @hadighasemi and made changes to wizard dialogue text based on this conversation. However, am waiting on @hyeukryu's return to the GA office on 11/7 to clarify IF process (specifically how/whether displaced persons are calculated).

###cc
@Charlotte-Morgan