Fix test failure on master builds
The handling of geojson files has changed in qgis master, causing
fields to be loaded again after commit the layer's changes. But
because there's no values present in the new field, there's nothing
in the resultant geojson file to maintain this newly added field,
and it's lost when reloading from disk.